Move migrate logic to dojo-world
commit-id:b188b831
