+++ crates/dojo-world/src/migration/mod.rs
@@ -1,2 +1,3 @@
 pub mod object;
+pub mod strategy;
 pub mod world;
