+++ crates/dojo-world/src/migration/strategy_test.rs
@@ -1,9 +1,9 @@
 use camino::Utf8PathBuf;
 use dojo_test_utils::sequencer::Sequencer;
-use dojo_world::config::{EnvironmentConfig, WorldConfig};
-use dojo_world::migration::world::WorldDiff;
 
-use crate::ops::migrate::prepare_for_migration;
+use crate::config::{EnvironmentConfig, WorldConfig};
+use crate::migration::strategy::prepare_for_migration;
+use crate::migration::world::WorldDiff;
 
 #[tokio::test]
 async fn test_migration() {
