+++ crates/sozo/Cargo.toml
@@ -36,9 +36,6 @@ url = "2.2.2"
 [dev-dependencies]
 assert_fs = "1.0.10"
 snapbox = "0.4.6"
-katana-core = { path = "../katana-core" }
-katana-rpc = { path = "../katana-rpc" }
-dojo-test-utils = { path = "../dojo-test-utils" }
 tokio = { version = "1.28.0", features = [ "full" ] }
 
 [[bin]]
