+++ crates/sozo/src/commands/migrate.rs
@@ -4,6 +4,7 @@ use anyhow::{anyhow, Result};
 use camino::Utf8PathBuf;
 use clap::Args;
 use dojo_world::config::{EnvironmentConfig, WorldConfig};
+use dojo_world::migration::strategy::prepare_for_migration;
 use dojo_world::migration::world::WorldDiff;
 use dotenv::dotenv;
 use scarb::core::Config;
@@ -11,7 +12,6 @@ use scarb::ops;
 use scarb::ui::Verbosity;
 
 use super::build::{self, BuildArgs, ProfileSpec};
-use crate::ops::migrate::prepare_for_migration;
 
 #[derive(Args)]
 pub struct MigrateArgs {
