--- crates/sozo/src/ops/mod.rs
@@ -1,1 +1,0 @@
-pub mod migrate;
